Crate problem084

Problem 84 (Monopoly odds)

Solution detail

An approximate answer will do! The exact probabilities can be calculated using Markov chain techniques, but since we are only interested in the relative order of the squares, not the exact probabilities, some Monte Carlo simulation will suffice.

That means that we just need to code up the rules of movement, simulate running around the board for a large (10^6 or 10^7) number of turns, and count the squares that we landed on the most.

Structs

MonopolyBoard

MonopolySimulator

Roller

Enums

StepResult

Functions

answer
Solve the problem, returning the answer as a String

main

solve
Find the three most common squares in a simulation using a dice with the given number of sides.