Struct problem084::MonopolySimulator

struct MonopolySimulator<'a> {
    board: MonopolyBoard<'a>,
    dice_roller: Roller,
    card_roller: Roller,
    doubles: usize,
    current_position: usize,
    counts: HashMap<usize, usize>,
}

Fields
board: MonopolyBoard<'a>dice_roller: Rollercard_roller: Rollerdoubles: usizecurrent_position: usizecounts: HashMap<usize, usize>
Methods

impl<'a> MonopolySimulator<'a>

pub fn new(dice_size: usize) -> MonopolySimulator<'a>

pub fn step(&mut self)

Simulate the result of a single roll

pub fn squares_by_popularity(&self) -> Vec<usize>

fn handle_doubles(&mut self, roll: (usize, usize)) -> StepResult

Handle tracking doubles as part of a turn, and go to JAIL if necessary

fn handle_moving(&mut self, roll: (usize, usize)) -> StepResult

Handle actually moving the piece as a result of the roll

fn handle_go_to_jail(&mut self) -> StepResult

Handle landing on Go To Jail

fn handle_chance(&mut self) -> StepResult

fn handle_community_chest(&mut self) -> StepResult